fix: split style-dictionary scripts
